% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{np_has_npub}
\alias{np_has_npub}
\title{Is \code{npub} available?}
\usage{
np_has_npub()
}
\value{
\code{TRUE} when \code{\link[=np_npub_bin]{np_npub_bin()}} resolves to an existing file.
}
\description{
Is \code{npub} available?
}
\examples{
np_has_npub()
}
