% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{np_table_rep}
\alias{np_table_rep}
\title{Table representation of a result field}
\usage{
np_table_rep(asset)
}
\arguments{
\item{asset}{The id of the asset holding values in field-domain order.}
}
\value{
A plain \code{list(kind = "table", asset)} representation.
}
\description{
Table representation of a result field
}
\examples{
np_table_rep("parcel-values")
}
