% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{np_volume_rep}
\alias{np_volume_rep}
\title{Volume representation of a result field}
\usage{
np_volume_rep(asset)
}
\arguments{
\item{asset}{The id of the asset holding the volume.}
}
\value{
A plain list \code{list(kind = "volume", asset)}.
}
\description{
Volume representation of a result field
}
\examples{
np_volume_rep("speech-t")
}
